+++ test/io.jl
@@ -1,6 +1,6 @@
 using Test: @test, @testset
 using ZebrafishHMM2023: ZebrafishHMM_G3, ZebrafishHMM_G4, ZebrafishHMM_TN03, ZebrafishHMM_TN04,
-    ZebrafishHMM_TN3, ZebrafishHMM_TN4,
+    ZebrafishHMM_TN3, ZebrafishHMM_TN4, ZebrafishHMM_G3_Sym, ZebrafishHMM_G4_Sym,
     save_hmm, load_hmm
 using Distributions: Normal, Gamma
 using Statistics: mean, std
@@ -79,6 +79,30 @@ end
     @test hmm_loaded.transition_matrix == hmm.transition_matrix
     @test mean(hmm_loaded.forw) == mean(hmm.forw)
     @test std(hmm_loaded.forw) == std(hmm.forw)
+    @test mean(hmm_loaded.turn) == mean(hmm.turn)
+    @test std(hmm_loaded.turn) == std(hmm.turn)
+end
+
+@testset "ZebrafishHMM_G3_Sym I/O" begin
+    h5_file = tempname()
+    hmm = ZebrafishHMM_G3_Sym(rand(), rand(3,3), 10.5, Gamma(0.6, 32))
+    save_hmm(h5_file, hmm)
+    hmm_loaded = load_hmm(h5_file, ZebrafishHMM_G3_Sym)
+    @test hmm_loaded.pinit_turn == hmm.pinit_turn
+    @test hmm_loaded.transition_matrix == hmm.transition_matrix
+    @test hmm_loaded.σforw == hmm.σforw
+    @test mean(hmm_loaded.turn) == mean(hmm.turn)
+    @test std(hmm_loaded.turn) == std(hmm.turn)
+end
+
+@testset "ZebrafishHMM_G4_Sym I/O" begin
+    h5_file = tempname()
+    hmm = ZebrafishHMM_G4_Sym(rand(), rand(4,4), 10.3, Gamma(0.6, 32))
+    save_hmm(h5_file, hmm)
+    hmm_loaded = load_hmm(h5_file, ZebrafishHMM_G4_Sym)
+    @test hmm_loaded.pinit_turn == hmm.pinit_turn
+    @test hmm_loaded.transition_matrix == hmm.transition_matrix
+    @test hmm_loaded.σforw == hmm.σforw
     @test mean(hmm_loaded.turn) == mean(hmm.turn)
     @test std(hmm_loaded.turn) == std(hmm.turn)
 end